Modularise AccInterface to V1 (basic), V2 (includes parseConfig function)
